fix(ci): python-full-coverage — extract merge-method helpers so merge_loop.py stays under the 1500-line limit
`test_first_party_code_files_stay_under_line_limit` failed because commit
317521973 (CI-start grace applied to the pre-merge recheck) grew
`pr_monitor_runner/merge_loop.py` from 1486 to 1511 lines, past the
MAX_FIRST_PARTY_FILE_LINES (1500) maintainability guard.

The #656 grace fix is correct and stays; this only restores the file under
the limit by moving the cohesive merge-method policy/classification/messaging
helpers verbatim into a new sibling module
`pr_monitor_runner/merge_methods.py` (the established decomposition pattern in
this package). `merge_loop.py` imports back the six names it still references;
the merge-method classifier test now imports the three pure helpers from their
new home. No behavior change — the move is coverage-neutral (same lines, same
exercising tests). merge_loop.py is now 1406 lines.
